Restructured the project so it can be used as a library in other programs as well
