+++ .gitignore
@@ -1,2 +1,3 @@
 imgcat
+!cmd/imgcat
 dist/
